Fixes to text, addition of Octave callout.
This commit fixes some sections of the text where the wrong function
was referred to (kelvin_to_celsius instead of fahr_to_celsius).

It adds a FIXME note as there's an image referred to but missing.

It adds a callout about Octave to confirm that Octave behaves like
MATLAB when searching for functions.

It fixes a couple of small typos.

It fixes some of the markdown to ensure that code sections render
correctly in one of the callouts.