I have made some corrections, and have a suggestion for slide 5.
Slide 02 - scripts:
print -dpng 'average.png'
This syntax doesn't appear in the help documentation for MATLAB R2015aSP1. Perhaps it is old syntax (certainly it works) but maybe best to use documented syntax to avoid confusion?

04 example a):
Missing empty string - code as it is produces an error.

05 functions:
Repeated code removed.

Suggestion:
Center function only works for vectors - the matrix example given in the lesson doesn't work, but it does work for the inflammation data. I've not had chance to rewrite the function to accept matrices, though a workaround would be to change the test to
z=zeros(2,1);center(z, 3)